backend/requirements.txt
astroid==2.1.0
autopep8==1.4.3
colorama==0.4.1
Django==2.1.4
djangorestframework==3.9.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pep8==1.7.1
pycodestyle==2.4.0
pylint==2.2.2
pytz==2018.7
six==1.12.0
wrapt==1.10.11
